requirements.txt
discord
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
flask
